Trading-Bot/requirements.txt
numpy
pandas
pyzmq
protobuf
grpcio-tools
